add simple fake endpoints for testing

+++ src/routes/history/+server.ts
@@ -1,6 +1,7 @@
 import { json, type RequestEvent } from '@sveltejs/kit';
 import type { Conversation } from '$lib/models/Contracts.js';
 import DatabaseService from '$lib/services/database-service';
+import { constants } from '$lib/constants';
 
 export async function POST({ request }: RequestEvent): Promise<Response> {
   const databaseService = await new DatabaseService().init();
@@ -17,6 +18,10 @@ export async function PUT({ request }: RequestEvent): Promise<Response> {
 export async function DELETE({ request }: RequestEvent): Promise<Response> {
   const databaseService = await new DatabaseService().init();
   const conversation = (await request.json()) as Conversation;
-  await databaseService.deleteHistory(conversation);
-  return new Response(undefined, { status: 204 });
+  const requestingUser = request.headers.get(constants.principalId) ?? constants.testUserId;
+  if (conversation.userId === requestingUser) {
+    await databaseService.deleteHistory(conversation);
+    return new Response(undefined, { status: 204 });
+  }
+  return new Response(undefined, { status: 401 });
 }
